pyproject.toml
[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"

[project]
name = "onshape_to_robot"
version = "1.7.9"
description = "Converting Onshape assembly to robot definition (URDF, SDF, MuJoCo) through Onshape API"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "requests",
    "commentjson",
    "colorama>=0.4.6",
    "numpy-stl",
    "transforms3d",
    "python-dotenv",
]

[project.urls]
Homepage = "https://onshape-to-robot.readthedocs.io/"
Repository = "https://github.com/rhoban/onshape-to-robot/"

[project.scripts]
onshape-to-robot = "onshape_to_robot.cli:main"
onshape-to-robot-bullet = "onshape_to_robot:bullet.main"
onshape-to-robot-mujoco = "onshape_to_robot:mujoco.main"
onshape-to-robot-clear-cache = "onshape_to_robot:clear_cache.main"
onshape-to-robot-edit-shape = "onshape_to_robot:edit_shape.main"
onshape-to-robot-pure-sketch = "onshape_to_robot:pure_sketch.main"

[project.optional-dependencies]
pymeshlab = ["pymeshlab"]
pybullet = ["pybullet"]
mujoco = ["mujoco"]

[tool.uv.build-backend]
module-root = ""
